back/requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aiosmtplib==2.0.2
asgiref==3.7.2
bcrypt==4.0.1
cryptography==41.0.6
email-validator==2.0.0
fast_captcha==0.2.1
fastapi==0.108.0
fastapi-limiter==0.1.6
fastapi-pagination==0.12.8
gunicorn==21.2.0
loguru==0.7.0
passlib==1.7.4
path==16.7.1
pre-commit==3.2.2
pydantic==2.2.1
pydantic-Settings==2.0.3
python-jose==3.3.0
python-multipart==0.0.6
redis[hiredis]==5.0.1
supervisor==4.2.5
tortoise-orm[asyncmy]==0.23.0
uvicorn[standard]==0.23.2
wait-for-it==2.2.2
